Set explicit type

+++ examples/AllowListWithAmount.sol
@@ -10,8 +10,8 @@ import "@franknft.eth/erc721-f/contracts/utils/AllowListWithAmount.sol";
  */
 contract AllowListWithAmountExample is ERC721FCOMMON, AllowListWithAmount {
     uint256 public constant MAX_TOKENS = 10000;
-    uint public constant MAX_PURCHASE = 31;
-    uint public tokenPrice = 1 ether;
+    uint256 public constant MAX_PURCHASE = 31;
+    uint256 public tokenPrice = 1 ether;
     bool public preSaleIsActive;
     bool public saleIsActive;
